+++ src/main/java/no/ntnu/idatt1002/data/User.java
@@ -6,14 +6,14 @@ import java.util.Date;
 
 public final class User {
 
-    private final int id;
+    private final long id;
     private final String username;
     private final String password;
     private final Date registerDate;
     private final long phoneNumber;
     private final Budget budget;
 
-    public User(int id, String username, String password, Date registerDate, long phoneNumber, Budget budget) {
+    public User(long id, String username, String password, Date registerDate, long phoneNumber, Budget budget) {
         if(username == null || username.isBlank())
             throw new IllegalArgumentException("Username cannot be null or blank");
         if(password == null || password.isBlank())
@@ -32,7 +32,7 @@ public final class User {
         this.budget = budget;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
